Move Azure client funcs to client.go
